tables.rb: Generate UTF-8 strings when possible.
Strings in `tables.rb` for the most part are common occurences.
MIME types and extension names are extremely likely to be referenced
in other gems and in the host application.

As such it's preferable to define them in the default encoding
so that the same interned string can be re-used.

The exception being the magic strings, they need to be binary.
For them we generate them with the `.b.freeze` suffix.